+++ src/Controller/Admin/UsersController.php
@@ -18,6 +18,7 @@ namespace MeCms\Controller\Admin;
 use Cake\Event\EventInterface;
 use Cake\Http\Response;
 use Cake\Mailer\MailerAwareTrait;
+use Laminas\Diactoros\UploadedFile;
 use Symfony\Component\Finder\Finder;
 use Thumber\Cake\Utility\ThumbManager;
 
@@ -247,10 +248,14 @@ class UsersController extends AppController
      * Changes the user's picture
      * @return void
      * @throws \Tools\Exception\ObjectWrongInstanceException|\ErrorException
+     * @todo `$UploadedFile` should always be a `\Laminas\Diactoros\UploadedFile` instance (see test)
      */
     public function changePicture(): void
     {
-        if ($this->getRequest()->getData('file')) {
+        /** @var ?\Laminas\Diactoros\UploadedFile $UploadedFile */
+        $UploadedFile = $this->getRequest()->getData('file');
+
+        if ($UploadedFile) {
             $id = $this->Auth->user('id');
 
             //Deletes any picture that already exists
@@ -258,9 +263,10 @@ class UsersController extends AppController
                 @unlink($file->getPathname());
             }
 
-            $filename = $id . '.' . pathinfo($this->getRequest()->getData('file')['tmp_name'], PATHINFO_EXTENSION);
+            $extension = pathinfo($UploadedFile instanceof UploadedFile ? $UploadedFile->getClientFilename() : $UploadedFile['tmp_name'], PATHINFO_EXTENSION);
+            $filename = $id . '.' . $extension;
 
-            $uploaded = $this->Uploader->setFile($this->getRequest()->getData('file'))
+            $uploaded = $this->Uploader->setFile($UploadedFile)
                 ->mimetype('image')
                 ->save(USER_PICTURES, $filename);