+++ templates/Users/login.php
@@ -11,14 +11,15 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
+ *
+ * @var \MeCms\View\View\AppView $this
  */
 
-/** @var \MeCms\View\View $this */
 $this->assign('title', $title = __d('me_cms', 'Login'));
 ?>
 
 <div id="login">
-    <?= $this->Form->create(null) ?>
+    <?= $this->Form->create(null, ['validation' => false]) ?>
     <fieldset>
         <?php
         echo $this->Form->control('username', [