+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.30 branch
 ### 2.30.13
+* fixed a bug with uploading the user's profile picture;
 * added the `AppTable::getCacheNameWithAssociated()` method. The `getCacheName()` method will now only return the cache
   name of the current table (string);
 * simplified the `Plugin` class;