+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 2.x branch
 ## 2.30 branch
 ### 2.30.13
+* fixed a bug for the "read more" button of posts and pages;
 * fixed a bug with uploading the user's profile picture;
+* fixed some small template bugs;
 * added the `AppTable::getCacheNameWithAssociated()` method. The `getCacheName()` method will now only return the cache
   name of the current table (string);
 * simplified the `Plugin` class;